rename modules for clarity, fix a warning
