+++ sdk-wasm/src/host.rs
@@ -1,16 +1,16 @@
 use http::uri::Authority;
 use iri_string::types::UriString;
+use kepler_lib::authorization::KeplerDelegation;
 use kepler_lib::cacaos::{
     siwe::{nonce::generate_nonce, Message, TimeStamp, Version},
     siwe_cacao::{SIWESignature, SiweCacao},
 };
 use kepler_lib::resource::OrbitId;
-use kepler_lib::zcap::KeplerDelegation;
 use serde::Deserialize;
 use serde_with::{serde_as, DisplayFromStr};
 use wasm_bindgen::prelude::*;
 
-use crate::zcap::DelegationHeaders;
+use crate::authorization::DelegationHeaders;
 
 #[serde_as]
 #[derive(Deserialize)]
